prov/gni: hookup of gni nic's with eps
Add code for associating underlying gni nics (the ones that
will be use for send/recv, rma, amos, etc.) with libfabric
ep's.

Add more fields to the gnix domain object for controlling
creation of underlying gni objects needed by the provider:
size of gni CQs attached to a gni nic, and the additional
cq mode bits that may be useful.

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>