prov/gni: add datagram subsystem
This commit adds support for an interface to the GNI datagram
(GNI_EpPostDataWId, etc). that can be used for exchanging the
initial SMSG data  needed to establish a GNI SMSG connection between two
GNI endpoints.  These connections will be used internally
in the GNI provider even in the case of FI_EP_RDM.

The datagram support is designed to be as self contained
and generic as possible.  When a process needs to send
information about a GNI smsg mailbox it takes the following
steps:

1) allocate a datagram with the gnix_dgram_alloc function.
   Since there are only a limited number of gni dgrams available
   (to avoid clogging up the kgni "session" state engine), the
   provider has to be prepared to handle a dgram not being available.
2) With an allocated datagram, fill in the dgram_in_buf buffer in
   the dgram with the info it wishes to exchange with the remote peer
3) Invoke the gnix_dgram_connect_post.  This results in the dgram
   being injected into the kgni "session" state engine, and ultimately
   delivery to the target node.  The intended receive process at the target
   must either post a matching gnix_dgram_connect_post or else have posted
   wildcard datagrams using the gnix_dgram_unbnd_post.
4) The process needs to poll on the EQ associated with the fabric object
   that was used to create the domain/ep's which require this internal
   GNI SMSG connection setup.  The data from the remote peer can be extracted
   from the dgram_out_buf in the datagram.

Note the EQ related code to implement 4 needs to be written.

The current plan is to have a pthread running in the gnix_dgram_prog_thread_fn.
One pthread per domain created.  The threads block in GNI_PostdataProbeWaitById.
When awoken with a completed datagram transaction, they'll inform the application
thread of the completion of the datagram exchange vi an EQ entry.

Fixes ofi-cray/libfabric-cray#60

Signed-off-by: Howard Pritchard <howardp@lanl.gov>