prov/gni: add nic counting
Use the ptag for tracking number of nics opened using a given ptag.
There are only up to GNI_PTAG_USER_END (253) possible ptags
so we can use a simple global array to track number of nics opened
for a given ptag.

The max # of nics/ptag will need to be determined using the number
of ranks in a job running on the node, as well as limits placed
by the job launcher on total number of FMA descriptors available
to the ranks in the job on the local node.

Keep the allocated nics in a global list to allow for multiple domains
with same ptag/cookie to use previously allocated nics.  Each domain
keeps a linked list of the domains it is using.

When the domain is closed, loop through the domain specific list of
gni nics being used, decrementing reference count.  If reference count
for a nic drops to zero, destroy the cdm, remove from the global list.

Start adding locks for thread safety support.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>