prov/gni: add a makefile for gni prov
There's been a request to be able to a different
layout of the gni provider source, a subdir for include
files, one for src files, and at some point, one for
sanity unit tests (not to replace fabtests).

Also move the gni part of the top level Makefile.am
into an included Makefile.  This allows us to make the
gni provider specific Makefile.am fancier without
needing to change the shared toplevel Makefile.am

Signed-off-by: Howard Pritchard <howardp@lanl.gov>