Add partial implementation of wait objects.
Added the skeleton for wait objects. A good portion of control operations are
implemented. This implementation cannot be used as wait objects since the
fi_wait function is not implemented. This necessitates more discussion on the
topic of progress. Once the progress model is added then fi_wait and the CQ
associated sread and sreadfrom can be implemented.

Signed-off-by: Ben Turrubiates <bturrubiates@lanl.gov>