Add generalized queue data structure. Closes #115.
Attempt to unify the CQ and EQ structures by creating a queue data structure.
The queue contains both an events list and a free list. Upon creation an
allocation and free function can be passed in and these will be used to
allocate objects for the free list and free objects in the destroy function.

Signed-off-by: Ben Turrubiates <bturrubiates@lanl.gov>