prov/gni: enhance gnix datagram for async prog
Enable a progress thread for handling incoming
gni datagrams.  The use of the progress thread
is determined by the control progress model
requested by the application.  The default for
gni provider is FI_PROGRESS_AUTO.

Code still needs to be added to support application
request for FI_PROGRESS_MANUAL.

This commit should solve a significant part of #97.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>