Add non-locking implementation of hash table
This commit adds a non-threadsafe implementation of the hash table.
To get a non-threadsafe hashtable, the ht_internal_locking variable
must be set to 0 in the gnix_hashtable_attr_t structure passed to
init. By default, ht_internal_locking is set to 0 if no attributes
are passed to init, which requires the external API to take into
account the following:

With a non-threadsafe hashtable (attr->ht_internal_locking == 0):
-- gnix_ht_init, gnix_ht_destroy, gnix_ht_insert, and gnix_ht_remove
   must be called under an exclusive lock. Each function modifies the
   internal structures in a non-atomic fashion.
-- gnix_ht_lookup can be called under a shared lock, or no lock at
   all so long as the other functions cannot be called while a lookup
   is in progress

When the threadsafe hashtable is in use, memory use of the structure
is increased linearly by the number of buckets in use. Each threadsafe
contains an embedded lock that increases the size of the bucket head
by the size of a pthread_rwlock_t. The non-threadsafe hashtable does
not have the embedded lock.

No changes need to be made to existing code that uses the hash table
as the API has not changed.

	modified:   prov/gni/include/gnix_hashtable.h
	modified:   prov/gni/src/gnix_hashtable.c
	modified:   prov/gni/test/hashtable.c