Fixed ffs and ffz functions for bitmaps of non-qword length multiples
If the bitmap has a length that is not a qword length multiple,
find_first_set_bit and find_first_zero_bit may return an index which
is greater than or equal to the length of the bitmap. This return value
is erroneous from the perspective of user, who should expect the following:

    0 <= value < bitmap->length

This commit fixes this behavior for the two mentioned functions and adds
two new tests to check for this behavior.

Signed-off-by: James Swaro <jswaro@cray.com>

	modified:   prov/gni/src/gnix_bitmap.c
	modified:   prov/gni/test/bitmap.c