prov/gni: implement counter class
This work implements the counter class functionality
for the gni provider.  Note the gni provider
currently does not claim the FI_RMA_EVENT
capability, so FI_REMOTE_WRITE and FI_REMOTE_READ
flags are not supported when binding a counter
to an ep.

Wait object functionality is also not fully implemented
yet.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>