prov/gni: set next field to NULL in slist element
Turns out the slist matching routine can get
easily messed up if next field in the element
being removed is invalid.  This can easily happen
if there's only a single element in the list,
or the match is made to the last element in the list.

Added a gni provider specific slist_insert_tail function
which sets the next field of the element being inserted
to NULL.  This is extra protection for problems found
in other work where bogus entries were dequeued off of the
list because the previous last element in the list had a bogus
value in its next field.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>