Fix assertion error in gnix_domain_close()
Because the memory registration cache clean up can be the last
reference to a nic, it must call _gnix_nic_free() rather than simply
decrementing the nic's ref count.  Not doing so resulted in a negative
ref count in gnix_domain_close(), because it was not removed from the
domain's nic list when the ref count reached 0.  This case is
possible, because the endpoint that allocated the nic can be (and
often is) closed before the domain.

Note that I did not similarly change the ref count decrement in
fi_gnix_mr_close(), because it seems to assume that it was not holding
the last reference to the domain.  I'm not sure if this is correct,
especially in the case of eager deregistration.  Probably we should
enable setting of the deregistration policy and test.

As part of this commit, I renamed the nic's list field to dom_nic_list
for clarity (and fixed the documentation for it and gnix_nic_list).

This gets rid of all the assertions in gnix_domain_close() for gnitest.

Fixes #259

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>