prov/gni: refactor the message path
Refactor the message path in preparation for addition
of code for an eager get path as well as rendezvous
path.

Separate out the tag match for incoming messages
into a separate function which can be reused
for all message paths.

A subsequent refactor will be needed for the
receive side in order to handle completion of
data transfers for paths other than the SMSG only
path.

This commit relates to issue #236

Signed-off-by: Howard Pritchard <howardp@lanl.gov>