prov/gni: introduced nic alloc options
This is a followup to the recently introduced
option for bound ep's: FI_EP_RDM types where the
user has specified the cdm id for the associated
gnix_cm_nic.

This PR enhances the ways a gnix_nic can be allocated.
If the app has requested a bound ep, rather than
consuming additional CQ's for a gnix nic, this option
allows the cdm/nic handles from the bound ep's cm_nic
to be reused when instantiating a gnix_nic for the ep.

This PR is the first in a step to allocate a single
gnix_cm_nic/domain in the case where the app isn't allocating
bound eps.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>