prov/gni: fix race condition with pending_vcs list
There was a problem revealed by the vc_management/vc_accept.
To fix the problem, the pending_vcs list on the nic had to
be switched to a dlist.  During the _gnix_vc_destroy call
a check is now made to see if the vc is in the pending_vcs
list and if so is removed.

Fixes #369

Signed-off-by: Howard Pritchard <howardp@lanl.gov>