prov/gni: refactor connection setup code
This is a major refactor of the connection
setup path within the GNI provider.  The
underlying driver for this refactor is issue

In the original GNI provider design, each
FI_EP_RDM endpoint had its own gnix_cm_nic
for connection management.  The problem with
this scheme is that after a few 10s or so of
EP rdms are created, a GNI CQ resource limit
is reached.

In this refactor, unless the FI_EP_RDM has a user
specified cdm_id, the EP rdms associated with a
given domain share the same gnix_cm_nic.

In order to do this, the datagram usage had to
be significantly changed.  Now connections are managed
by a two-stage datagram exchange approach.  Wildcard
datagrams no longer have mailboxes associated with them,
they are purely used to sink incoming connection requests
and responses.

This refactor also uncovered some bugs in other parts
of the connection management path.

Some of the criterion tests had to be modified owing
to GNI provider internal changes.

Fixes #230.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>