prov/gni: av refactor
add enough info in av entry to handle multiplexing
EP RDM types over a single gnix_cm_nic.

add a hash table for FI_AV_MAP entries

add an entry in the vc struct to hold the address
of the cm_nic that the vc needs to post a datagram
to for mailbox set.

Fix a problem with the fi_cancel of receives
that shows up now that we are actually caching
gnix_address info for FI_AV_MAP type.

Fixes #384
Fixes #339

prov/gni: enhance av critierion test

make sure that we test the table resize
code, at least some.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>