Add 2 new MR tests
These tests repeatedly register the same buffer, each time for a
different size.  One test increases the size, the other decreases it.
The one that increases the size is an explicit version of what is
currently done in the rdm_sr::send_autoreg_uncached under the covers
to automatically register the source buffer.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>