Fix for MR cache flush.
The red-black tree match function returns the "best" match in a
(potentially re-balanced) tree, which may not be the exact match if
one exists.  This may cause a problem in the stale tree when flushing
via LRU, because the LRU entry may match an entry in the tree that is
not an exact match (which is should be).  This results a different
entry being removed from the LRU list than from the stale tree.

This removes the "best" match from the stale tree.  If this match is
not the LRU entry, then remove the match from the LRU list, and put
the original LRU entry back on the LRU list.

The inuse tree can also run into a similar problem during
deregistration which is not being addressed here.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>