Add a test that does pipelined tsend/trecv.
This new test (multi_tsend_trecv_pipelined) is a combination of
rdm_tagged_sr/multi_tsend_trecv and rdm_rx_overrun/all_to_one.  It
only works with manual progress if the mbox max credit limit is
increased, so I restructured the setup function so we can call the
domain open_ops function.  The new tests runs in a new suite
rdm_tagged_sr_progress_manual.

For the restructure, I split up the set up function so that we could
set the progress model and also set the mbox max credit limit.  All
the tests are currently using FI_PROGRESS_MANUAL, but that should be
changed when FI_PROGRESS_AUTO is working (see comments for where).

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>