prov/gni: fix problems for osu_bibw
The osu_bibw was bringing out an issue with
scheduling of vc's for processing.  Namely,
a vc should be scheduled if its gotten
SMSG notifications on its RX CQ for a mailbox.
This could indicate a return of SMSG credits
which allows for progressing of pending GNI
SMSG sends.

Also, Schedule the receive complete message back
to sender in a request send_fn method to
allow for replay, rather than as a tx completer
function, which is not a suitable place for
transactions that may fail due to lack of
SMSG credits, etc.

With these changes, the osu_bibw now passes.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>