Update the rdm_rx_overrun test (still doesn't work)
The previous assertion in this test went away with some of the recent
changes.  When I enabled this test, I noticed it wasn't going through
the overrun path as intended, since we were using the same domain for
all endpoints and hence always making progress.  So I rewrote it to
use separate domains for each endpoint and ran into an issue in
__gnix_vc_connect_to_self that eventually leads to a Criterion
assertion.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>