Fix bug in gnix_fab_req usage and do a bit of cleanup
The gnix_fab_req structures are allocated out of a freelist, and the
field are not zero'ed between uses.  This commit zeros common fields
(tx_failures and modes) in _gnix_fr_alloc(), and removes similar
initialization near call sites.  The tx_failure field also needs to be
reset when reusing the gnix_fab_req for the rendezvous message.

While in there, I also removed the unused retries field.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>