Update rdm_rx_overrun test to actually exercise the RX overrun path.
Howard got the basic test working with a bug fix, and now I finally
got this test to exercise the RX overrun path.

The test sets the RX CQ size to 1 and then sends messages from many
endpoints to endpoint 0.  Note that even tho we request an RX CQ size
of 1, the ugni internal minimum appears to be something like 511, so
we have to send more than 1 message per endpoint.  Currently 60
endpoints send 10 message each before endpoint 0 queries its CQ.

When run with FI_LOG_LEVEL=warn, we see the warning due to calling
__nic_rx_overrun.

TODO: Figure out a good way to make sure we continue to test the
overrun path, without having to manually check if we get the warning.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>