prov/gni: check for GNI_VERSION_FMA_CHAIN_TRANS..
The GNI provider is making use of chained FMA post
operations, a feature than is only available in
CLE 5.2UP04 and newer.  Check for the presence
of a sentinel macro GNI_VERSION_FMA_CHAIN_TRANSACTIONS
in gni_pub.h. Configure GNI provider off if this
sentinel is not found in gni_pub.h

Fixes #553

Signed-off-by: Howard Pritchard <howardp@lanl.gov>