prov/gni:config - add war for a bug in gni
Well it turns out that gni_pub.h isn't entirely
accurate about GNI_CtPostFma.  It turns out some
of the functionality was implemented in CLE 5.2UP03,
but not what is currently needed by the GNI provider.
So the sentinel check GNI_VERSION_FMA_CHAIN_TRANSACTIONS
only partly works, screening out CLE5.2UP02 and older
but not the problem with CLE 5.2UP03.

Add a hack workaround to check for this CLE release and
if that's what installed, don't currently build the GNI
provider.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>