prov/gni: fix an issue with datagram exchange
Turns out there was a difficult to hit edge case where
a vc could end up sending a datagram both for connect
request and connect request response(ack), resulting in
an error in connection setup at the target.

The problem was insufficient checking in the ack
progress function.  It needs to check whether or not
the vc has either already moved to connected state or
a GNI datagram for this vc setup has already been posted.

Add some additional GNIX_WARN's for various edge cases.
The warning statements helped in debugging this problem.

Fixes ofi-cray/libfabric-cray#599

Signed-off-by: Howard Pritchard <howardp@lanl.gov>