prov/gni: refactor cm nic code
Refactor the cm nic code as part of a campaign
to reduce use of Aries NIC resources per process in
preparation for KNL.

This refactor now has the cm_nic embed a gnix_nic
struct.  The gnix_nic can then be reused for
RDMA network requests.  This also cuts down on
the number of HW CQs required per process.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>