Modify progress functions to exit after any VC reschedules
The only way to really guarantee we break out of the progress loop is
to return when any VC is rescheduled.  Just checking that we don't see
the first VC again is not enough, as it could be another VC that is
repeatedly rescheduled.  If we don't break out of the progress loop,
we run into a potential deadlock situation when 2 or more endpoints
are trying to connect at the same time.

While in there I renamed __gnix_nic_vc_* functions to __gnix_vc_nic_*
to be consistent with our naming convention and changed a few
assert(0) calls to call GNIX_FATAL so that we could get better info
when we hit them.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>