prov/gni: match getinfo man page better
The GNI provider wasn't appending the secondary capabilities
in the returned caps field in the fi_getinfo call.  This was
leading to some confusion on the part of MPICH users/developers
as to the capabilities that the GNI provider supported.

Follow the pattern used by the sockets provider and have
separate macros for primary and secondary caps, and or in the
secondary caps in the returned fi_info struct.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>