prov/gni: Map VCs in EP by fi_addr
Previously, VCs were mapped in an EP by their gnix_address.  With this, When
looking up a VC for a TX operation, an AV look up is done to translate fi_addr
to gnix address, followed by a EP lookup to translate gnix address to VC.  VCs
will now be mapped by fi_addr, allowing us to skip the AV lookup during a TX
operation.  It is possible for a connection request to be received from a peer
that has not been inserted into the local AV.  In this case, the mapping from
gnix address to fi_addr is unknown at the time of VC creation.  An
'unmapped_vc' list in the EP is used to track these VCs until they are mapped
using the AV, at which time they can be inserted into the EP's VC look up
table.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>