prov/gni: add XPMEM on-node transfer method
This commit adds support for use of XPMEM for
on-node transfers for GNI provider's send/recv
rendezvous protocol

The feature by default will be automatically
enabled when building on a Cray XC system running
CLE 5.2up04 or higher and using a C11 compliant C
compiler.

The feature can be disabled at configure time with
the

--disable-xpmem

configure option.

If the provider is configured with XPMEM support
(the default) the feature can be partially disabled
by setting the

GNIX_DISABLE_XPMEM

environment variable.  This will disable use of XPMEM
for inter process memcpy, but will not disable the
initial setup and teardown of XPMEM related setup
steps.

disable xpmem bypass for criterion test

Signed-off-by: Howard Pritchard <howardp@lanl.gov>