prov/gni: fix a possible race condition with dgrams
When an app is using manual progress for control, but
has data progress set to auto, its possible for multiple
threads to be progressing datagrams.  This can result in
occasions where multiple threads may detect the presence
of a complete datagram transaction via GNI_PostDataProbeById,
but only one of the threads will actually dequeue the datagram
in the subsequent call to GNI_EpPostDataTestById, the others
will get an GNI_RC_NO_MATCH error return.

This was not being properly handled.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>