Add missing symbols and includes to fi_ext_gni.h
The include for 'bool' was missing, as was the
actual definitions for the enum gnix_fab_req_type.

To resolve the issue, the header was included at the top of
the file and an alternate enumeration was provided. The alternate
enumeration is used when redefining the values internally so
as to prevent enumeration value mismatches.

Signed-off-by: James Swaro <jswaro@cray.com>