prov/gni: have stx open allocate a gnix_nic
Some applications might want to use the stx approach
to bind EP's to different "hardware" nic resources.
With this commit, gnix_fid_stx is no longer a no-op
but a way for an app to allocate an underlying
gnix_nic struct and to bind it to one more EP's
allocated from the same domain.

This is closer to the intent of the OFI libfabric
API for shared TX structs than the original GNI
provider code.

The original way of allocating a gnix_nic during
the gnix_ep_open phase is unchanged unless the
info argument's ep_attr->tx_ctx_cnt field is set
to FI_SHARED_CONTEXT.

Note that the fabtest which checks the functionality
of tx shared contexts has a similar issue to that found
for the scalable endpoint test.

A fix to the fabtest should be merged in to that repo
before merging this PR in to libfabric-cray.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>
(cherry picked from commit d2cec4df0b30d7a5303b72d829542f9010bebff3)