prov/gni: avoid initialization via constructor
Try to avoid run-time initialization of data structures in src/gnix_init.c
to avoid issues with the (undefined) order in which library constructors
get called during process startup.  Initialize global atomics in gnix_init()
when GNI_INI is invoked to determine if it is an available provider in
addition to in the gnix_init() constructor in case it gets called before the
constructor runs.

Signed-off-by: Steven Vormwald <sdvormwa@cray.com>