prov/gni: use a free list for VC's
Turns out some of the criterion tests which intentionally
do things to check for errors, etc. have a tendency to
bring out some issues in the destruction of objects within
the GNI provider when EPs, domains, etc. are closed and
there are outstanding transactions/GNI RX CQEs etc. still
outstanding.  As a partial step to addressing this implement
a nic-based free list of VC's that aren't completely destructed
till the nic is destructed.

Also be more careful when processing GNI RX CQEs.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>