prov/gni: Fix kdreg tests
We didn't notice that the notifier tests stopped being tested with
kdreg when we changed the configure script to use separate flags for
compiling gnitest and the provider.  This commit adds the kdreg flags
to the gnitest flags and fixes the notifier tests to use the new
notifier interface from bc3f605.

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>