prov/gni: help out KNL with symbol relocation
Turns out the KNL branch predictor wigdet has a hard
time with branches to symbols in shared libraries that
are located more than 4 GB of VM space away from where
the jump is going to be initiated.  This leads to
successively worse performance for software that traverses,
say, multiple shared libraries to access the network.

This problem is decribed in the Intel Xeon Phi Processor
High Performance Programming: Knights Landing Edition.
In that text, a workaround using the LD_PREFER_MAP_32BIT_EXEC
environment variable is described, but unfortunately that
required glibc 2.23 or higher, which is not yet available on
Cray XC systems.

As a way to help out the KNL with the GNI provider, add
a configury option that allows for the libfabric.so to
be statically linked against the Cray ugni library.

This commit adds in a

--enable-ugni-static

configury option which allows for static linking against
libugni.a when building the libfabric with GNI provider.

Note that an ldd of the resulting libfabric.so will still
show a dependency on libugni.so but that is owing to a dependency
of the ALPS libalpslli.so on libugni.so and does not impact
the static linking of the GNI provider/libfabric.so against libugni.a.

A GNI provider micro benchmark shows the advantage of static
linking of the libfabric.so against libugni.a:

With static linking:

```
~/cray_tests_install/bin>aprun -n 2 -N 1 ./rdm_pingpong
1 threads
1                         4.82                4.82                4.82
2                         4.86                4.86                4.86
4                         4.84                4.84                4.84
8                         4.87                4.87                4.87
16                        4.84                4.84                4.84
32                        4.89                4.89                4.89
64                        4.91                4.91                4.91
128                       4.94                4.94                4.94
256                       5.04                5.04                5.04
512                       5.32                5.32                5.32
1024                      6.36                6.36                6.36

```

with default dynamic linking:

```
1                         5.06                5.06                5.06
2                         5.08                5.08                5.08
4                         5.06                5.06                5.06
8                         5.09                5.09                5.09
16                        5.07                5.07                5.07
32                        5.06                5.06                5.06
64                        5.12                5.12                5.12
128                       5.16                5.16                5.16
256                       5.22                5.22                5.22
512                       5.49                5.49                5.49
1024                      6.60                6.60                6.60
```

This configury option is not enabled by default for now as it might confuse
GNI provider developers trying to debug problems using modified
ugni libraries.

Signed-off-by: Howard Pritchard <hppritcha@gmail.com>