prov/gni: adjust mbox_allocator criterion test
Turns out that on production systems like NERSC edison,
it can be difficult to assemble large (128 MB and greater)
page sizes reliably.  This was causing periodic failure
of criterion test runs, and thus spurious failures in
CI testing.

This commit adds a "may_fail" argument to the mbox allocator
routine in the test.  If the mbox allocator returns -FI_ENOMEM,
don't treat as an error for these large page size requests.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>