prov/gni: fix returned caps field in info struct
MPICH/CH4 testing revealed a bug in the caps field
in the returned info struct from fi_getinfo.  It
turned out that if the user had not supplied a
hints argument, GNI provider would just return 0 for
capabilities.  This got the MPICH/CH4 confused and
it would also devolve to the sockets provider,
or abort if the gni provider was explicitely selected.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>